+++ backend/auth/auth.js
@@ -158,7 +158,7 @@ var buildActivity = async function(profile){
       topics = topicResponse.data.filter(item => item.parent_id);
       
       const uploadIds = topics.map( (item) => {
-          if ( (item) && (item.name) && (String(item.name).indexOf("repo") === -1) && (String(item.name).indexOf("branch") ===-1) && (String(item.name).indexOf("varClass") === -1)){
+          if ( (item) && (item.name) && (String(item.name).indexOf("repo") === -1) && (String(item.name).indexOf("branch") ===-1) && (String(item.name).indexOf("varClass") === -1) && (String(item.name).indexOf("project") === -1)){
               return item.name
           }
           return ""