+++ frontend/src/changelog.md
@@ -3,13 +3,17 @@
 <br />
 
 ## Version 2.0.2
-July 10 2023
+July 11 2023
  - Added support for any amount of schema types provided is still the first class citizen
+ - Added projects, users (by email) can be assigned to a project which has a name and a set of datasets/editions
+   - Users added to the project will be able to see those datasets and editions, but not discussion related to the datasets/editions (comments are visible)
+ - Fixed a bug with filtering by highlight on files and fields
+
 
  <br />
 
 ## Table Of Contents
-- [Version 2.0.2 (July 10 2023)](#version-202)
+- [Version 2.0.2 (July 11 2023)](#version-202)
 - [Version 2.0.1 (June 28 2023)](#version-201)
 - [Version 2.0.0 (June 22 2023)](#version-200)
 - [Version 1.7.1 (June 2 2023)](#version-171)